Brings the 92 commits development has gained since the last sync: lite
nodes and snapshots, stratum mining, the Tx PoW server, P2P v19, the
seed/peer discovery rework and the attached console.

Six files conflicted. Beyond the mechanical resolution, two seams needed
real fixes so the merged behaviour matches both sides:

NetNode.cpp - development's rewritten resolve_seed_nodes always added the
built-in SEED_NODES. test-mnodes had made the built-ins conditional so a
--network profile's seed list replaces them rather than adding to them;
without that a testnet node dials the mainnet seeds. The built-in loop is
now gated on the configured seed list being empty.

DaemonCommandsHandler.cpp - development moved command output to out(), so
an attached console can capture it, and replaced setHandler with
register_command. The four masternode commands still wrote to std::cout
and were registered the old way, so they compiled but returned nothing
over /console. They are re-registered through register_command and print
through out(); background-thread notices stay on std::cout.

Verified: full Release build clean, masternodetest 370 checks 0 failures,
cryptotest clean, and one Wrkzd carries both --lite/--import-lite-snapshot
and --network/--mn-signing-key.

+++ src/utilities/FormatTools.cpp
@@ -20,7 +20,7 @@
 
 namespace Utilities
 {
-    std::string get_mining_speed(const uint64_t hashrate)
+    std::string get_mining_speed(const double hashrate)
     {
         std::stringstream stream;
